refactor: CaughtHttpExchange 패키지 변경
