+++ api/src/main/kotlin/com/backgu/amaker/api/common/annotation/CaughtHttpExchange.kt
@@ -1,4 +1,4 @@
-package com.backgu.amaker.api.config
+package com.backgu.amaker.api.common.annotation
 
 import org.springframework.web.service.annotation.HttpExchange
